Updated to camelCase TF calls.
Split the functionality into detector (which detects the object) and node/action who listen to point clouds and request detection. The action gets the pose as the goal and detects the nearest object instance to the goal. The node detects all observed instances of the object.



git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@18465 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: ce5acbe8b5d9f4e7b152ae4902d5f63f2734fcf2

+++ sandbox/labeled_object_detector/src/pcd_misc.cpp
@@ -142,7 +142,7 @@ void pcd_misc::cluster_pcd_points(const robot_msgs::PointCloud& centers,
   unsigned int iPtOut=0;
   for(unsigned int iPt=0;iPt<num_pts;iPt++)
   {
-    if(cluster_ids[iPt]==iPt)
+    if(cluster_ids[iPt]==int(iPt))
     {
       cluster_ids_final[iPt]=iPtOut;
       
@@ -151,7 +151,7 @@ void pcd_misc::cluster_pcd_points(const robot_msgs::PointCloud& centers,
   }
   for(unsigned int iPt=0;iPt<num_pts;iPt++)
   {
-    if(cluster_ids[iPt]!=iPt)
+    if(cluster_ids[iPt]!=(int)iPt)
     {
       int clusterOutID=cluster_ids_final[cluster_ids[iPt]];
       cluster_ids_final[iPt]=clusterOutID;
@@ -177,7 +177,7 @@ void pcd_misc::cluster_ids_to_cluster_indices(const std::vector<int>& cluster_id
     unsigned int num_in_cluster=0;
     for(unsigned int iPt=0;iPt<num_pts;iPt++)
     {
-      if(cluster_ids[iPt]==iC)
+      if(cluster_ids[iPt]==(int)iC)
       {
         num_in_cluster++;
       }
@@ -187,7 +187,7 @@ void pcd_misc::cluster_ids_to_cluster_indices(const std::vector<int>& cluster_id
     unsigned int out_location=0;
     for(unsigned int iPt=0;iPt<num_pts;iPt++)
     {
-      if(cluster_ids[iPt]==iC)
+      if(cluster_ids[iPt]==(int)iC)
       {
         cluster_members[out_location]=iPt;
         out_location++;
